refactor: move Imprint component to shared directory and update imports
