+++ src/app/app.routes.ts
@@ -1,6 +1,6 @@
 import { Routes } from '@angular/router';
 import { MainContentComponent } from './main-content/main-content.component';
-import { ImprintComponent } from './imprint/imprint.component';
+import { ImprintComponent } from './shared/imprint/imprint.component';
 
 export const routes: Routes = [
   { path: '', component: MainContentComponent },
